\docType{data}
\name{yan}
\alias{yan}
\title{Single cell RNA-Seq data extracted from a publication by Yan et al.}
\format{An list of 3 objects. raw: \code{data.frame} with 20286 rows and 90 columns, before log2 transformation; sc_cnt: \code{data.frame} with 20286 rows and 90 columns, after log2 transformation; sc_label: \code{character} cell labels provided by authors of the original publication.}
\source{
\url{http://dx.doi.org/10.1038/nsmb.2660}
For gene expression dataframe, columns represent cells, rows represent genes expression values.
}
\usage{
data(yan)
}
\description{
Single cell RNA-Seq data extracted from a publication by Yan et al.
}
\keyword{datasets}




